added if statement for deleting session.txt

+++ main.py
@@ -3,7 +3,8 @@ import simplematrixbotlib as botlib
 from src.poll import Poll
 
 # deleting session.txt file, to prevent message looping
-os.remove("session.txt")
+if os.path.exists("session.txt"):
+    os.remove("session.txt")
 # Initialize bot credentials from environment variables
 creds = botlib.Creds(
     homeserver=os.environ['HOMESERVER'],
